rnndb: fix VARYING_COMPONENT_USE enum naming
The current naming is very confusing, as the most commonly used value
is named UNUSED.

However the value 0 is simply used for generic components that are not
affected by any of the rasterizer settings. Other values are used for
color varyings (which are affected by rasterizer flatshade and provoking
vertex setting) and pointcoord, which is a rasterizer generated varying.