rnndb: rename RA depth HDEPTH_DISABLE to FORWARD_Z
Apparently this bits gates passing Z from RA to SH. The blob enables
this bit when the shader reads the fragcoord.z value as seen in
dEQP-GLES2.functional.shaders.builtin_variable.fragcoord_xyz

This bit must also be enabled when the late Z stage is used, as the
depth value has no way to reach the late Z stage when it isn't passed
through to SH, even if SH doesn't touch the value at all.

I mistook this bit as being hdepth related when I first encountered it
in tests using the late Z stage on GPUs with the RA_WRITES_DEPTH feature,
as without a proper Z value reaching the late depth stage the only depth
related thing that works is the coarse primitive based depth rejection
in RA, so the result of disabling the bit looked somewhat coarse/hdepth
related.